merge: fix/be/meerkat-39 -> develop

See merge request s08-webmobile1-sub2/S08P12B107!73

+++ server/src/main/java/B107/server/meerkat/entity/ExpiredMarker.java
@@ -37,8 +37,8 @@ public class ExpiredMarker extends BaseAtTime implements Serializable {
     @JoinColumn(name = "member_idx")
     private Member member;
 
-    private Float lat; // 위도
-    private Float lng; // 경도
+    private Double lat; // 위도
+    private Double lng; // 경도
     private String location; // 위치 정보
 
     @Column(name = "reg_date")
