+++ server/src/main/java/B107/server/meerkat/entity/Marker.java
@@ -38,8 +38,8 @@ public class Marker extends BaseAtTime implements Serializable {
     @JoinColumn(name = "member_idx")
     private Member member;
 
-    private Float lat; // 위도
-    private Float lng; // 경도
+    private Double lat; // 위도
+    private Double lng; // 경도
     private String location; // 위치 정보
 
     @JsonDeserialize(using = LocalDateTimeDeserializer.class)
